% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VersionHistory.R
\name{VersionHistory}
\alias{VersionHistory}
\title{Version history table}
\usage{
VersionHistory(c("date", "Version"), c("date", "Version"), ...)
}
\arguments{
\item{...}{Pass 1 list per version, with a date (dd/mm/yyyy) and a comment. The most recent update should be given first. For example: 'c("29/09/2022", "Updated with more content"),c("01/09/2022", "Initial release")'}
}
\description{
Function to create a simple version history table that can be added to documents
}
\details{
This function is used for our guidance documents. It may have limited use cases beyond those.
}
\author{
Ben Cuff
}
